-Started..UI Design -Started..Business Logic Design

+++ app/src/main/java/com/erdemserhat/a1orwin/model/Lack.kt
@@ -9,8 +9,8 @@ public class Lack {
 
         }
 
-        fun rollDice(): Int {
-            return (Math.random() * 7).toInt() + 1
+        fun rollDice(): Float {
+            return ((Math.random() * 7).toInt() + 1).toFloat()
         }
 
     }
